feat: change allocation percentage to decimal
Update allocation percentage type from INTEGER to 
NUMERIC(5, 2) across schema, models, and UI to 
support fractional values (e.g., 45.50%).

+++ src/main/java/br/gov/sisgos/dto/AlocacaoResponseDTO.java
@@ -19,7 +19,7 @@ public class AlocacaoResponseDTO {
     private String itemContratacao;
     private String mesReferencia;
     private String nomeProfissional;
-    private Integer percentualAlocacao;
+    private BigDecimal percentualAlocacao;
     private String documentoReferencia;
     private BigDecimal custoMensalPerfil;
     private BigDecimal custoAlocacao;
